+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed empty semantic node for dropdown menu items in ios [#4996](https://github.com/rokwire/illinois-app/issues/4996).
 - Show images of the sporting events in the events feed [#5533](https://github.com/rokwire/illinois-app/issues/5533).
+- Make Event widget cards a fixed height [#5507](https://github.com/rokwire/illinois-app/issues/5507).
 ### Added
 - Added "All Payment Types" to Payment Type dropdown in Map2 and Dining2HomePanel [#5352](https://github.com/rokwire/illinois-app/issues/5352).